refactor(config): make editor the mutation boundary
