+++ canfar/models/config.py
@@ -17,8 +17,6 @@ from pydantic import (
 )
 
 if TYPE_CHECKING:
-    from collections.abc import Iterable
-
     from pydantic.fields import FieldInfo
 
 from pydantic_settings import (
@@ -31,8 +29,6 @@ from pydantic_settings.sources import EnvSettingsSource
 
 from canfar import CONFIG_PATH
 from canfar.config.editor import ConfigurationEditor as _ConfigurationEditor
-from canfar.config.editor import get_value as _get_value
-from canfar.config.editor import set_value as _set_value
 from canfar.models.active import ActiveConfig
 from canfar.models.auth import (
     AuthenticationCredential,
@@ -190,41 +186,9 @@ class Configuration(BaseSettings):
             file_secret_settings,
         )
 
-    def _heal_default_storage(self) -> None:
-        """Restore default Storage Identifiers on Servers saved by an older client.
-
-        Older clients keyed a discovered VOSpace Service by its Server Name, so
-        an existing configuration holds ``canfar`` instead of ``arc`` and never
-        gained later defaults such as ``vault``. Healing is scoped to the
-        default Servers so federated Servers keep their Server Name keys.
-
-        This normalizes the loaded Configuration in memory only; the file on
-        disk is rewritten when something independently calls ``save()``.
-        """
-        for name, default in default_servers.items():
-            server = self.servers.get(name)
-            if server is None or not default.storage or server.idp != default.idp:
-                continue
-            storage = dict(server.storage)
-            legacy = storage.pop(name, None)
-            if legacy is None and storage:
-                # Deliberate Storage Identifiers are configuration, not stale defaults.
-                continue
-            if legacy is not None:
-                leaf = str(legacy.uri).rpartition("/")[2] or name
-                storage.setdefault(leaf, legacy)
-            for identifier, service in default.storage.items():
-                storage.setdefault(identifier, service.model_copy(deep=True))
-            if storage != server.storage:
-                self.servers[name] = server.model_copy(
-                    update={"storage": storage},
-                    deep=True,
-                )
-
     @model_validator(mode="after")
     def _normalize_and_validate_servers(self) -> Configuration:
         """Inject Server Names and validate Server and Storage Identifier keys."""
-        self._heal_default_storage()
         updated: dict[str, Server] = {}
         server_by_identifier: dict[str, str] = {}
         for name, server in self.servers.items():
@@ -314,293 +278,10 @@ class Configuration(BaseSettings):
 
         return self
 
-    def save(self) -> None:
-        """Save the current configuration to the default YAML file."""
-        from canfar.config.store import save_config  # noqa: PLC0415
-
-        save_config(self)
-
     @property
     def editor(self) -> _ConfigurationEditor:
         """Return the bound editing and persistence boundary."""
         return _ConfigurationEditor(self)
-
-    def _validated_copy(self, **updates: Any) -> Configuration:
-        """Validate a source-isolated copy of this complete Configuration."""
-        data = {**self.model_dump(mode="python"), **updates}
-        candidate = self.__class__.model_construct()
-        self.__class__.__pydantic_validator__.validate_python(
-            data,
-            self_instance=candidate,
-        )
-        return candidate
-
-    def _replace_state(
-        self,
-        *,
-        active: ActiveConfig | None = None,
-        authentication: dict[str, AuthenticationCredential] | None = None,
-        servers: dict[str, Server] | None = None,
-    ) -> None:
-        """Validate and install a complete Authentication and Server state."""
-        # Validate only this candidate; BaseSettings construction would reload and
-        # merge persisted/environment sources, resurrecting keys being removed.
-        candidate = self._validated_copy(
-            active=self.active if active is None else active,
-            authentication=(
-                self.authentication if authentication is None else authentication
-            ),
-            servers=self.servers if servers is None else servers,
-        )
-        self.active = candidate.active
-        self.authentication = candidate.authentication
-        self.servers = candidate.servers
-
-    def get_value(self, path: str) -> Any:
-        """Get a nested configuration value via dotted path (e.g. 'console.width')."""
-        return _get_value(self, path)
-
-    def set_value(self, path: str, value: Any) -> Configuration:
-        """Return a new validated Configuration with a dotted-path value updated."""
-        return _set_value(self, path, value)
-
-    def get_credential(self, idp: str) -> AuthenticationCredential:
-        """Return the saved authentication credential for an IDP key.
-
-        Args:
-            idp: Canonical identity provider key.
-
-        Returns:
-            Matching authentication credential.
-
-        Raises:
-            KeyError: If no credential exists for ``idp``.
-        """
-        if idp not in self.authentication:
-            msg = f"Authentication record for IDP '{idp}' not found."
-            raise KeyError(msg)
-        return self.authentication[idp]
-
-    def _get_server_by_name(self, name: str) -> Server:
-        """Return a known server by Server Name."""
-        if name not in self.servers:
-            msg = f"Server '{name}' not found."
-            raise KeyError(msg)
-        return self.servers[name]
-
-    def upsert_credential(self, credential: AuthenticationCredential) -> None:
-        """Insert or replace a validated Authentication Record.
-
-        Args:
-            credential: Authentication Record to store by its IDP key.
-        """
-        self._replace_state(
-            authentication={**self.authentication, credential.idp: credential},
-        )
-
-    def update_credential(self, credential: AuthenticationCredential) -> None:
-        """Replace an existing validated Authentication Record.
-
-        Raises:
-            KeyError: If no Authentication Record exists for the credential IDP.
-        """
-        self.get_credential(credential.idp)
-        self.upsert_credential(credential)
-
-    def set_active_authentication(self, idp: str) -> None:
-        """Select an Authentication Record and its remembered Server, if any."""
-        self.get_credential(idp)
-        remembered = self.get_remembered_server_for_idp(idp)
-        if remembered is not None:
-            self.set_active_selection(idp, remembered)
-            return
-
-        selections = self._server_selection_history()
-        server_name = self.active.server
-        if server_name is not None:
-            try:
-                active_server = self.get_active_server()
-            except KeyError:
-                server_name = None
-            else:
-                if active_server.idp != idp:
-                    server_name = None
-        self._replace_state(
-            active=self.active.model_copy(
-                update={
-                    "authentication": idp,
-                    "server": server_name,
-                    "servers": selections,
-                },
-            ),
-        )
-
-    def remove_authentication(self, idp: str) -> None:
-        """Remove an Authentication Record and its Science Platform Servers."""
-        authentication = dict(self.authentication)
-        authentication.pop(idp, None)
-        servers = {
-            name: server for name, server in self.servers.items() if server.idp != idp
-        }
-        selections = {
-            selected_idp: name
-            for selected_idp, name in self.active.servers.items()
-            if selected_idp != idp
-        }
-
-        if not authentication:
-            self.purge_authentication()
-            return
-
-        active = self.active.model_copy(update={"servers": selections})
-        if active.authentication == idp:
-            active = active.model_copy(
-                update={
-                    "authentication": next(iter(authentication)),
-                    "server": None,
-                },
-            )
-        self._replace_state(
-            active=active,
-            authentication=authentication,
-            servers=servers,
-        )
-
-    def purge_authentication(self) -> None:
-        """Reset Authentication and Server state while preserving other settings."""
-        self._replace_state(
-            active=default_active.model_copy(deep=True),
-            authentication={
-                key: credential.model_copy(deep=True)
-                for key, credential in default_authentication.items()
-            },
-            servers={
-                name: server.model_copy(deep=True)
-                for name, server in default_servers.items()
-            },
-        )
-
-    def get_server_by_uri(self, uri: str | AnyUrl) -> Server:
-        """Return a known server by IVOA URI.
-
-        Args:
-            uri: Server URI to resolve.
-
-        Returns:
-            Matching server record.
-
-        Raises:
-            KeyError: If no server exists for ``uri``.
-        """
-        target = str(uri)
-        for server in self.servers.values():
-            if server.uri is not None and str(server.uri) == target:
-                return server
-        msg = f"Server '{target}' not found."
-        raise KeyError(msg)
-
-    def get_active_server(self) -> Server:
-        """Return the active science platform server record.
-
-        Raises:
-            KeyError: If no active server is selected.
-        """
-        if self.active.server is None:
-            msg = "No active server selected."
-            raise KeyError(msg)
-        return self._get_server_by_name(self.active.server)
-
-    def get_server_for_idp(self, idp: str) -> Server:
-        """Return the best-known server for an IDP.
-
-        Uses the active server when it matches ``idp``; otherwise returns the
-        first saved server for the IDP.
-
-        Args:
-            idp: Canonical identity provider key.
-
-        Returns:
-            Server record for the IDP.
-
-        Raises:
-            KeyError: If no server exists for ``idp``.
-        """
-        if self.active.authentication == idp and self.active.server is not None:
-            return self.get_active_server()
-
-        for server in self.servers.values():
-            if server.idp == idp:
-                return server
-        msg = f"No server found for IDP '{idp}'."
-        raise KeyError(msg)
-
-    def get_remembered_server_for_idp(self, idp: str) -> Server | None:
-        """Return the last selected server for ``idp`` when still valid.
-
-        Args:
-            idp: Canonical identity provider key.
-
-        Returns:
-            Matching server record, or ``None`` when no remembered selection is
-            available for ``idp``.
-        """
-        name = self._server_selection_history().get(idp)
-        if name is None or name not in self.servers:
-            return None
-        server = self.servers[name]
-        if server.idp != idp:
-            return None
-        return server
-
-    def _server_selection_history(self) -> dict[str, str]:
-        """Return remembered selections seeded with the current active pair."""
-        selections = dict(self.active.servers)
-        active_name = self.active.server
-        if active_name is None or active_name not in self.servers:
-            return selections
-        server = self.servers[active_name]
-        if server.idp == self.active.authentication and server.name is not None:
-            selections[self.active.authentication] = server.name
-        return selections
-
-    def set_active_selection(self, idp: str, server: Server) -> None:
-        """Persist ``idp`` and ``server`` as the active pair.
-
-        Args:
-            idp: Canonical identity provider key.
-            server: Server record to activate.
-
-        Raises:
-            ValueError: If the server has no Server Name.
-        """
-        if server.name is None:
-            msg = "Server name is required for active selection."
-            raise ValueError(msg)
-
-        selected = server.model_copy(update={"idp": idp}, deep=True)
-        servers = {**self.servers, server.name: selected}
-        selections = self._server_selection_history()
-        selections[idp] = server.name
-        active = self.active.model_copy(
-            update={
-                "authentication": idp,
-                "server": server.name,
-                "servers": selections,
-            },
-        )
-        self._replace_state(active=active, servers=servers)
-
-    def upsert_server(self, server: Server) -> None:
-        """Insert or replace a validated server record keyed by Server Name."""
-        self.upsert_servers((server,))
-
-    def upsert_servers(self, servers: Iterable[Server]) -> None:
-        """Insert or replace validated server records in one state change."""
-        updated = dict(self.servers)
-        for server in servers:
-            if server.name is not None:
-                updated[server.name] = server
-        self._replace_state(servers=updated)
 
 
 __all__ = ["CONFIG_PATH", "Configuration", "ConsoleConfig"]
